+++ pyvela/pyvela/pyvela_script.py
@@ -18,8 +18,8 @@ from pyvela.results import SPNTAResults
 def parse_args(argv):
     parser = ArgumentParser(
         prog="pyvela",
-        description="A command line interface for the Vela.jl pulsar timing &"
-        " noise analysis package. Uses emcee for sampling. This may not be "
+        description="A command line interface for the Vela.jl pulsar timing & "
+        "noise analysis package. Uses emcee for sampling. This may not be "
         "appropriate for more complex datasets. Write your own scripts for "
         "such cases.",
         formatter_class=ArgumentDefaultsHelpFormatter,
@@ -44,11 +44,6 @@ def parse_args(argv):
         "-P",
         "--prior_file",
         help="A JSON file containing the prior distributions for each free parameter. (Ignored if `-J` option is used.)",
-    )
-    parser.add_argument(
-        "--no_marg_gp_noise",
-        action="store_true",
-        help="Don't analytically marginalize the correlated Gaussian noise amplitudes.",
     )
     parser.add_argument(
         "-A",
@@ -237,7 +232,7 @@ def main(argv=None):
             args.tim_file,
             cheat_prior_scale=args.cheat_prior_scale,
             custom_priors=(args.prior_file if args.prior_file is not None else {}),
-            marginalize_gp_noise=(not args.no_marg_gp_noise),
+            marginalize_gp_noise=True,
             analytic_marginalized_params=args.analytic_marg,
             center_epochs=args.center_epochs,
         )